backend/pubmed-search-tester-extract-events/requirements.txt
functions-framework==3.*
Flask==3.1.0
google-genai
vertexai==1.71.1